Added new player model
Changed the contraints on the player model to prevent player from moving on its own along the x and z axis. Apparently the cube model was doing this too, but the mesh collider on the new model made it worse.

Changed moving from rb.MovePosition to just setting the transform of the object.