Dragon Collider Mesh Updated GAME BREAKING BUG
I updated the dragon's primitive mesh, and enabled physics for its rigidbody.

Player and enemy rigidbodies should have x and z rotation constraints so that physics can still affect them but we can use not so perfect colliders.

I need to work on fixing a bug that is caused whenever using the scar L gun in the inventory system and calling set active. The katana works fine.