Client - Server Player Controller Rpc Calls
I changed the logic so that all movement and look calculations that used to be done in player controller are done in the server script attached to the server's network manager instance.

The client calls an rpc to the server which provides the input, which action to perform, and which client to perform it on.

It doesn't propagate back to the clients yet